Refactor SinglePost to allow replies to be made to main post - Create PostDetails component which displays details of post - Wrap ReplyFormContextProvider around PostDetails in SinglePost - Refactor Reply, Edit, and Delete buttons into PostOptions component
